Move VM trap stack helpers into core
