+++ docs/internal/testing/checklists/architecture-improvements.md
@@ -59,6 +59,7 @@ Issue #51 init benchmark baseline evidence (2026-04-28):
 - [x] `ARCH-RTCORE-23` Move bytecode error, reader, and alignment helper ownership into `trust-runtime-core`, preserve `trust-runtime::bytecode` compatibility modules for existing host code, and validate core bytecode helper behavior plus bytecode container/section/validation integration tests, clippy, full-map doctor, and diagram drift gates.
 - [x] `ARCH-RTCORE-24` Split host runtime-to-bytecode construction into `trust-runtime::bytecode::build_module_from_runtime*` functions, move runtime and harness call sites off the `BytecodeModule::from_runtime*` inherent API, and keep the old methods as temporary compatibility shims before moving the bytecode container type.
 - [x] `ARCH-RTCORE-25` Move pure bytecode format records and raw format constants into `trust-runtime-core`, keep `BytecodeModule` host-side until the public compatibility shims are retired, and validate core format discriminants plus runtime bytecode container, metadata, section decode, and validation tests.
+- [x] `ARCH-RTCORE-26` Start VM execution movement by classifying VM modules, moving the pure VM trap/error mapping and operand stack into `trust-runtime-core`, preserving runtime compatibility re-export modules, and validating core VM tests plus `bytecode_vm_core` and runtime VM lib filters.
 - [x] `ARCH-FULL-01` Add `full-architecture-refactor-program-checklist.md` as the umbrella board that states the runtime-core split is one branch of the full SOLID/KISS and zero-silent-bug program, not the whole program.
 - [x] `ARCH-FULL-02` Add `architecture-doctor-full-map-execution-checklist.md` so `architecture-doctor --full-map` becomes a concrete prerequisite for enforcing workspace, runtime-core, CLI, host-surface, KISS, API, and diagram rules.
 - [x] `ARCH-FULL-03` Add `hir-mutation-hardening-execution-checklist.md` for the F2 silent-bug gap: `symbol_import`, `type_check::const_eval`, and aggregate initializer validation mutation survivors.
